+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#90](https://github.com/dblock/jenkins-ansicolor-plugin/pull/90): Added support for italic, strikeout, framed and overline attributes - [@Joe-Merten](https://github.com/Joe-Merten).
 * [#92](https://github.com/dblock/jenkins-ansicolor-plugin/pull/92): Fixed high intensity colors for both foreground and background - [@Joe-Merten](https://github.com/Joe-Merten).
 * [#94](https://github.com/dblock/jenkins-ansicolor-plugin/pull/94): Added support for negative (inverse colors) attribute - [@Joe-Merten](https://github.com/Joe-Merten).
+* [#96](https://github.com/dblock/jenkins-ansicolor-plugin/pull/96): Added support for xterm 256 colors and 24 bit colors - [@Joe-Merten](https://github.com/Joe-Merten).
 * Your contribution here.
 
 0.4.3 (11/20/2016)